requirements.txt
aiofiles==0.4.0
aiohttp_requests==0.1.3
disnake==2.4.0
numpy==1.21.6
pandas==1.3.4
python-dotenv==0.20.0
